fix: update notarization script
