+++ frontend/app/package.json
@@ -16,7 +16,7 @@
     "cypress:open": "cypress open",
     "cypress:run": "cypress run --headed",
     "electron:build": "pnpm run build && pnpm run electron:package",
-    "electron:package": "electron-builder build --config .electron-builder.config.cjs --publish onTag",
+    "electron:package": "electron-builder build --config .electron-builder.config.ts --publish onTag",
     "electron:serve": "tsx scripts/serve.ts",
     "link:verify": "tsx scripts/check-external-links.ts",
     "lint:style": "stylelint 'src/**/*.{vue,scss}'",
@@ -72,7 +72,7 @@
     "zod": "3.24.2"
   },
   "devDependencies": {
-    "@electron/notarize": "3.0.0",
+    "@electron/notarize": "3.0.1",
     "@intlify/unplugin-vue-i18n": "6.0.5",
     "@pinia/testing": "1.0.0",
     "@tsconfig/node22": "22.0.0",
